Unify implementations of DRKey fetcher
