+++ go/lib/drkey/fetcher/fetcher.go
@@ -1,28 +1,46 @@
-package fetcher
+// Copyright 2020 ETH Zurich
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package grpc
 
 import (
 	"context"
 
 	"github.com/scionproto/scion/go/lib/addr"
-	drkeyctrl "github.com/scionproto/scion/go/lib/ctrl/drkey"
+	ctrl "github.com/scionproto/scion/go/lib/ctrl/drkey"
 	"github.com/scionproto/scion/go/lib/drkey"
 	"github.com/scionproto/scion/go/lib/serrors"
-	libgrpc "github.com/scionproto/scion/go/pkg/grpc"
+	sd_drkey "github.com/scionproto/scion/go/pkg/daemon/drkey"
+	sc_grpc "github.com/scionproto/scion/go/pkg/grpc"
 	cppb "github.com/scionproto/scion/go/pkg/proto/control_plane"
 )
 
+// DRKeyFetcher obtains end-host key from the local CS.
 type Fetcher struct {
-	Dialer libgrpc.Dialer
+	Dialer sc_grpc.Dialer
 }
 
-func (f *Fetcher) Lvl1(ctx context.Context, meta drkey.Lvl1Meta) (drkey.Lvl1Key, error) {
+var _ sd_drkey.Fetcher = (*Fetcher)(nil)
+
+func (f *Fetcher) Lvl1Key(ctx context.Context, meta drkey.Lvl1Meta) (drkey.Lvl1Key, error) {
 	conn, err := f.Dialer.Dial(ctx, addr.SvcCS)
 	if err != nil {
 		return drkey.Lvl1Key{}, serrors.WrapStr("dialing", err)
 	}
 	defer conn.Close()
 	client := cppb.NewDRKeyIntraServiceClient(conn)
-	protoReq, err := drkeyctrl.IntraLvl1ToProtoRequest(meta)
+	protoReq, err := ctrl.IntraLvl1ToProtoRequest(meta)
 	if err != nil {
 		return drkey.Lvl1Key{},
 			serrors.WrapStr("parsing AS-AS request to protobuf", err)
@@ -31,78 +49,90 @@ func (f *Fetcher) Lvl1(ctx context.Context, meta drkey.Lvl1Meta) (drkey.Lvl1Key,
 	if err != nil {
 		return drkey.Lvl1Key{}, serrors.WrapStr("requesting AS-AS key", err)
 	}
-	key, err := drkeyctrl.GetASASKeyFromReply(meta, rep)
+	key, err := ctrl.GetASASKeyFromReply(meta, rep)
 	if err != nil {
 		return drkey.Lvl1Key{}, serrors.WrapStr("obtaining AS-AS key from reply", err)
 	}
 	return key, nil
 }
 
-func (f *Fetcher) ASHost(ctx context.Context, meta drkey.ASHostMeta) (drkey.ASHostKey, error) {
+func (f Fetcher) ASHostKey(ctx context.Context,
+	meta drkey.ASHostMeta) (drkey.ASHostKey, error) {
+
 	conn, err := f.Dialer.Dial(ctx, addr.SvcCS)
 	if err != nil {
 		return drkey.ASHostKey{}, serrors.WrapStr("dialing", err)
 	}
 	defer conn.Close()
 	client := cppb.NewDRKeyIntraServiceClient(conn)
-	protoReq, err := drkeyctrl.ASHostMetaToProtoRequest(meta)
+	protoReq, err := ctrl.ASHostMetaToProtoRequest(meta)
 	if err != nil {
 		return drkey.ASHostKey{},
-			serrors.WrapStr("parsing AS-Host request to protobuf", err)
+			serrors.WrapStr("parsing AS-HOST request to protobuf", err)
 	}
 	rep, err := client.ASHost(ctx, protoReq)
 	if err != nil {
-		return drkey.ASHostKey{}, serrors.WrapStr("requesting AS-Host key", err)
+		return drkey.ASHostKey{}, serrors.WrapStr("requesting AS-HOST key", err)
 	}
-	key, err := drkeyctrl.GetASHostKeyFromReply(rep, meta)
+
+	key, err := ctrl.GetASHostKeyFromReply(rep, meta)
 	if err != nil {
-		return drkey.ASHostKey{}, serrors.WrapStr("obtaining AS-Host key from reply", err)
+		return drkey.ASHostKey{}, serrors.WrapStr("obtaining AS-HOST key from reply", err)
 	}
+
 	return key, nil
 }
 
-func (f *Fetcher) HostAS(ctx context.Context, meta drkey.HostASMeta) (drkey.HostASKey, error) {
+func (f Fetcher) HostASKey(ctx context.Context,
+	meta drkey.HostASMeta) (drkey.HostASKey, error) {
+
 	conn, err := f.Dialer.Dial(ctx, addr.SvcCS)
 	if err != nil {
 		return drkey.HostASKey{}, serrors.WrapStr("dialing", err)
 	}
 	defer conn.Close()
 	client := cppb.NewDRKeyIntraServiceClient(conn)
-	protoReq, err := drkeyctrl.HostASMetaToProtoRequest(meta)
+	protoReq, err := ctrl.HostASMetaToProtoRequest(meta)
 	if err != nil {
 		return drkey.HostASKey{},
-			serrors.WrapStr("parsing Host-AS request to protobuf", err)
+			serrors.WrapStr("parsing HOST-AS request to protobuf", err)
 	}
 	rep, err := client.HostAS(ctx, protoReq)
 	if err != nil {
-		return drkey.HostASKey{}, serrors.WrapStr("requesting Host-AS key", err)
+		return drkey.HostASKey{}, serrors.WrapStr("requesting HOST-AS key", err)
 	}
-	key, err := drkeyctrl.GetHostASKeyFromReply(rep, meta)
+
+	key, err := ctrl.GetHostASKeyFromReply(rep, meta)
 	if err != nil {
-		return drkey.HostASKey{}, serrors.WrapStr("obtaining Host-AS key from reply", err)
+		return drkey.HostASKey{}, serrors.WrapStr("obtaining HOST-AS key from reply", err)
 	}
+
 	return key, nil
 }
 
-func (f *Fetcher) HostHost(ctx context.Context, meta drkey.HostHostMeta) (drkey.HostHostKey, error) {
+func (f Fetcher) HostHostKey(ctx context.Context,
+	meta drkey.HostHostMeta) (drkey.HostHostKey, error) {
+
 	conn, err := f.Dialer.Dial(ctx, addr.SvcCS)
 	if err != nil {
 		return drkey.HostHostKey{}, serrors.WrapStr("dialing", err)
 	}
 	defer conn.Close()
 	client := cppb.NewDRKeyIntraServiceClient(conn)
-	protoReq, err := drkeyctrl.HostHostMetaToProtoRequest(meta)
+	protoReq, err := ctrl.HostHostMetaToProtoRequest(meta)
 	if err != nil {
 		return drkey.HostHostKey{},
-			serrors.WrapStr("parsing Host-Host request to protobuf", err)
+			serrors.WrapStr("parsing HOST-AS request to protobuf", err)
 	}
 	rep, err := client.HostHost(ctx, protoReq)
 	if err != nil {
 		return drkey.HostHostKey{}, serrors.WrapStr("requesting Host-Host key", err)
 	}
-	key, err := drkeyctrl.GetHostHostKeyFromReply(rep, meta)
+
+	key, err := ctrl.GetHostHostKeyFromReply(rep, meta)
 	if err != nil {
 		return drkey.HostHostKey{}, serrors.WrapStr("obtaining Host-Host key from reply", err)
 	}
+
 	return key, nil
 }
