--- go/pkg/daemon/drkey/grpc/BUILD.bazel
@@ -1,36 +1,0 @@
-load("//lint:go.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["fetcher.go"],
-    importpath = "github.com/scionproto/scion/go/pkg/daemon/drkey/grpc",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//go/lib/addr:go_default_library",
-        "//go/lib/ctrl/drkey:go_default_library",
-        "//go/lib/drkey:go_default_library",
-        "//go/lib/serrors:go_default_library",
-        "//go/pkg/daemon/drkey:go_default_library",
-        "//go/pkg/grpc:go_default_library",
-        "//go/pkg/proto/control_plane:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["fetching_test.go"],
-    deps = [
-        ":go_default_library",
-        "//go/lib/drkey:go_default_library",
-        "//go/lib/xtest:go_default_library",
-        "//go/pkg/grpc/mock_grpc:go_default_library",
-        "//go/pkg/proto/control_plane:go_default_library",
-        "//go/pkg/proto/control_plane/mock_control_plane:go_default_library",
-        "//go/pkg/proto/drkey:go_default_library",
-        "@com_github_golang_mock//gomock:go_default_library",
-        "@com_github_golang_protobuf//ptypes:go_default_library_gen",
-        "@com_github_stretchr_testify//require:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//test/bufconn:go_default_library",
-    ],
-)
